+++ src/main/java/ch/heigvd/pro/b04/sessions/SessionRepository.java
@@ -1,8 +1,13 @@
 package ch.heigvd.pro.b04.sessions;
 
+import ch.heigvd.pro.b04.moderators.Moderator;
+import ch.heigvd.pro.b04.polls.ServerPoll;
+import java.util.List;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SessionRepository extends JpaRepository<Session, SessionIdentifier> {
-  Optional<Session> findByCode(String code);
+public interface SessionRepository extends JpaRepository<ServerSession, SessionIdentifier> {
+  Optional<ServerSession> findByCode(String code);
+
+
 }
