+++ src/test/java/ch/heigvd/pro/b04/sessions/ServerSessionCodeTest.java
